Move error to top level
The parser errors are just standard errors and won't report what position the exception occurred on. By making the error type global, it can be used as a base type for the parser's errors.
